requirements.txt
# Visual Physics World-Model Requirements
# Target: Mac M4, 16GB RAM with MPS support

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pillow>=10.0.0

# Optional: For SSIM computation
scikit-image>=0.21.0

# Optional: For video saving
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Optional: For visualization
matplotlib>=3.7.0
tensorboard>=2.13.0
